+++ extensions/github-codespaces/src/open-codespace.tsx
@@ -27,7 +27,12 @@ export default function Command() {
   const handleSortByChange = setCriteria;
 
   if (!data?.codespaces) {
-    return <></>;
+    return (
+      <List
+        searchBarAccessory={<SortBy onSortByChange={handleSortByChange} />}
+        isLoading={true}
+      ></List>
+    );
   }
 
   const ListByCriteria = {