+++ extensions/github-codespaces/CHANGELOG.md
@@ -1,3 +1,9 @@
 # GitHub Codespaces Changelog
 
+## [Added menubar and stopping] - 2022-09-19
+
+- Added the ability to stop active codespaces
+- Added sorting/grouping to the codespace list
+- Implemented beta menubar command to show active codespaces
+
 ## [Initial Version] - 2022-09-11